Revert "Revert "General cleanup and prep for TailwindCSS (#33)""
This reverts commit 6b5b9c0152d5b869d3e5becc5c0c4c2a1d2dd96a.
