+++ src/components/BaseHead.astro
@@ -1,28 +1,19 @@
 ---
+import { smartypants } from 'smartypants';
+
 export interface Props {
-  title: string;
-  description: string;
-  canonicalURL: URL | string,
+  title?: string;
+  description?: string;
+  canonicalURL?: URL | string
   image?: string;
 }
-const { title, description, canonicalURL } = Astro.props;
+const { description, canonicalURL = Astro.request.canonicalURL } = Astro.props;
 const image = new URL(Astro.props.image || './social.png', Astro.site);
----
 
-<!-- Blocking script -->
-<script>
-  const marqueePaused = (localStorage.getItem('astro:marquee-paused') || 'false') === 'true';
-  window.addEventListener('DOMContentLoaded', () => {
-    const checkbox = document.querySelector('#marquee-pause');
-    if(checkbox) {
-      checkbox.checked = marqueePaused;
-      document.body.classList.add('js');
-    }
-  });
-  if ('paintWorklet' in CSS) {
-    CSS.paintWorklet.addModule('/pixel.worklet.js');
-  }
-</script>
+const title = [Astro.props.title, 'Astro']
+  .filter(Boolean)
+  .join(' | ')
+---
 
 <!-- Global Metadata -->
 <meta charset="utf-8" />
@@ -38,7 +29,7 @@ const image = new URL(Astro.props.image || './social.png', Astro.site);
 <link rel="preload" as="style" href="/styles/main.css">
 
 <!-- Primary Meta Tags -->
-<title>{title}</title>
+<title set:html={smartypants(title, 1)}></title>
 <meta name="title" content={title} />
 <meta name="description" content={description} />
 <link rel="canonical" href={canonicalURL}/>
