+++ src/components/Checklist.astro
@@ -2,6 +2,42 @@
 const { cols = 2 } = Astro.props;
 ---
 
-<ul class="checklist" style={`--cols: ${cols};`} role="list">
+<ul style={`--cols: ${cols};`} role="list">
     <slot />
 </ul>
+
+<style>
+    ul {
+        font-size: var(--size-500);
+        --marker-size: 3rem;
+        display: grid;
+        gap: 0.75rem;
+        margin: 0 auto;
+        padding: 0;
+        list-style: none;
+        grid-template-columns: repeat(1, minmax(0, 1fr));
+        justify-content: center;
+        width: 100%;
+        max-width: 72rem;
+        list-style: none;
+    }
+
+    @media (min-width: 24rem) {
+        ul {
+            grid-template-columns: repeat(2, minmax(0, 1fr));
+            justify-content: flex-start;
+        }
+    }
+
+    @media (min-width: 40rem) {
+        ul {
+            padding: 0 1.5rem;
+        }
+    }
+
+    @media (min-width: 52rem) {
+        ul {
+            grid-template-columns: repeat(var(--cols), minmax(0, 1fr));
+        }
+    }
+</style>
