+++ src/components/ChecklistItem.astro
@@ -5,12 +5,12 @@ const Wrapper = href ? 'a' : Fragment;
 const wrapperProps = href ? { href, target: '_blank', rel: 'noopener noreferrer' } : {};
 ---
 
-<li class="checklist-item" role="listitem">
+<li role="listitem">
     <Wrapper {...wrapperProps}>
-        <span class="checklist-marker">
+        <span class="marker">
             <slot name="icon" />
         </span>
-        <span class="checklist-text"><slot /></span>
+        <span class="text"><slot /></span>
         <Sprite aria-hidden="true" class="arrow" pack="mdi" name="arrow-right" size="1em" />
     </Wrapper>
 </li>
@@ -22,11 +22,11 @@ a {
     color: inherit;
     text-decoration: none;
 }
-.checklist-marker {
+.marker {
     transition: transform 200ms cubic-bezier(0.23, 1, 0.320, 1);
     transition-property: transform, background, color;
 }
-a:is(:hover, :focus) .checklist-marker {
+a:is(:hover, :focus) .marker {
     color: var(--color-dawn);
     background: var(--color-purple);
     transform: scale(1.10);
@@ -39,5 +39,24 @@ a .arrow {
 a:is(:hover, :focus) .arrow {
     transform: translateX(0.25em);
     opacity: 1;
+}
+.marker {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: var(--marker-size);
+    height: var(--marker-size);
+    background: var(--color-dawn);
+    color: var(--color-midnight);
+    border-radius: 50%;
+    margin-right: 1rem;
+}
+.text {
+    font-size: var(--size-600);
+}
+@media (min-width: 24rem) {
+    li {
+        margin-left: 2rem;
+    }
 }
 </style>
