+++ src/components/Nav.astro
@@ -38,3 +38,169 @@ const items = [
         ))}
     </ul>
 </nav>
+
+<script type="module" hoist>
+    import "/src/scripts/nav.ts";
+</script>
+
+<style>
+nav {
+    --offset-height: calc(var(--offset) * 1rem);
+    width: 100%;
+    position: sticky;
+    top: calc((var(--offset-height) * -1) - 1px);
+    margin-top: -2rem;
+    padding: 0.5rem 0;
+    background: linear-gradient(0deg, rgba(var(--color-tan-rgb), 0.7), rgba(var(--color-tan-rgb), 1));
+    -webkit-backdrop-filter: blur(10px);
+    backdrop-filter: blur(10px);
+    z-index: 2;
+}
+nav.margin-bottom {
+    margin-bottom: -4.5rem;
+}
+nav.invert {
+    background: transparent;
+    color: white;
+    -webkit-backdrop-filter: none;
+    backdrop-filter: none;
+}
+nav.invert a {
+    color: white;
+}
+
+nav::before {
+    content: '';
+    display: block;
+    width: 100%;
+    height: var(--offset-height);
+    z-index: -2;
+}
+nav.invert::before {
+    background: #1f1638;
+    margin-top: -0.5rem;
+}
+
+nav::after {
+    position: absolute;
+    content: '';
+    top: var(--offset-height);
+    right: 0;
+    bottom: 0;
+    left: 0;
+    display: flex;
+    box-shadow: var(--shadow-md);
+    opacity: 0;
+    transition-property: opacity;
+    transition-duration: 200ms;
+    transition-timing-function: cubic-bezier(0.23, 1, 0.320, 1);
+    z-index: -1;
+}
+
+nav[stuck]::after {
+    opacity: 0.5;
+    transition-duration: 600ms;
+}
+
+ul {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    font-size: 1.33rem;
+    padding: 0 0.5rem;
+}
+li {
+    color: var(--color-dusk);
+    font-family: var(--font-display);
+    font-weight: 400;
+    font-size: var(--size-500);
+}
+nav.invert li {
+    color: white;
+}
+li + li {
+    margin-left: 0;
+}
+[aria-current="page"] {
+    font-weight: 700;
+}
+.logo {
+    font-size: 1rem;
+    margin-right: auto;
+}
+.logo a {
+    display: flex;
+    align-items: center;
+    color: var(--color-midnight);
+}
+.logo a:hover,
+.logo a:focus {
+    color: var(--color-midnight);
+}
+.logo a:active {
+    color: var(--color-midnight);
+}
+.logo :global(svg) {
+    transform: scale(0.8);
+}
+.logomark {
+    color: inherit;
+}
+nav.invert .logo :global(svg) {
+    color: white;
+    fill: white;
+}
+.wordmark {
+    display: none;
+    color: inherit;
+    margin-top: -0.5rem;
+}
+.hidden-mobile {
+  display: none;
+}
+
+@media (min-width: 52rem) {
+    nav {
+        /* height: calc(2px + 4.5rem); */
+    }
+    .logo :global(svg) {
+        transform: scale(1);
+    }
+    .wordmark {
+        display: block;
+    }
+    .hidden-mobile {
+      display: block;
+    }
+}
+a {
+    display: flex;
+    height: 100%;
+    color: var(--color-midnight);
+    text-decoration: none;
+    transition: all 200ms cubic-bezier(0.165, 0.84, 0.44, 1);
+    transition-property: transform, color;
+    padding: 0 0.5rem;
+}
+a:focus,
+a:hover {
+    color: var(--color-purple);
+}
+a:active {
+    color: var(--color-blue);
+}
+
+.icon a :global(svg) {
+    transform: scale(0.9);
+}
+
+.icon a:focus :global(svg),
+.icon a:hover :global(svg) {
+    transform: scale(1);
+}
+
+[astro-icon="logo"] {
+    margin-right: 0.25em;
+    transform: translateY(0.5em);
+}
+</style>
