+++ src/components/Noscript.astro
@@ -1,0 +1,38 @@
+<noscript></noscript>
+
+<style>
+noscript {
+    order: -1;
+    height: 100vh;
+    width: 100vw;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-family: var(--font-display);
+    font-size: var(--size-600);
+    text-align: center;
+    color: var(--color-midnight);
+    background: linear-gradient(to bottom, var(--color-tan), white);
+}
+noscript::before {
+    --size: var(--size-1000);
+    padding-top: var(--size);
+    padding-left: 1rem;
+    padding-right: 1rem;
+    content: 'You need to enable JavaScript to view this site.';
+    background-size: var(--size);
+    background-position: center top;
+    background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 140 140'%3E%3Cg fill='%23514375'%3E%3Cpath d='M129.7 100v20h-10v-20z'/%3E%3Cpath d='M119.69 100h20v10h-20zM10 100v20h10v-20z'/%3E%3Cpath d='M20 100H0v10h20zm0-90H0v10h20z'/%3E%3Cpath d='M20 20V0H10v20zm79.69 30h20v10h-20zm-39.21 0h20v10h-20zM29.73 20h80v10h-80zm-9.97 50h99.93v10H19.76zm10 10h79.93v10H29.76zm-.03 10h20v10h-20zm30.75 0h20v10h-20zm29.21 0h20v10h-20zM19.73 30h100v10h-100zm0 10h100v10h-100zm0 20h100v10h-100zm0-10h20v10h-20zM120 10h20v10h-20z'/%3E%3Cpath d='M120 20V0h10v20z'/%3E%3C/g%3E%3C/svg%3E%0A");
+}
+noscript::after {
+    --size: var(--size-1000);
+    content: "Just kidding—you're going to love Astro!";
+    font-size: 0.75em;
+    padding-bottom: var(--size);
+    background-size: var(--size);
+    background-position: 50% 100%;
+    background-image: url("data:image/svg+xml,%3Csvg width='140' height='141' fill='none' xmlns='http://www.w3.org/2000/svg'%3E%3Cpath fill='%23af43ff' d='M85 80.325v-20h10v20zM45 80.325v-20h10v20zM65 40.325h10v60H65z'/%3E%3Cpath fill='%23af43ff' d='M55 80.325h30v10H55z'/%3E%3C/svg%3E");
+    animation: bounce 1s steps(4) alternate-reverse infinite;
+}
+</style>
