--- src/components/Scripts.astro
@@ -1,19 +1,0 @@
-{Astro.request.url.pathname === '/' && <noscript></noscript>}
-<script type="module" hoist>
-    import "../scripts/nav.ts";
-</script>
-<script type="module" src="/scripts/defer.js"></script>
-<script type="module">
-const checkbox = document.querySelector('#marquee-pause');
-if(checkbox) {
-    function onChange(element) {
-        const paused = element.checked;
-        localStorage.setItem('astro:marquee-paused', `${paused}`);
-    }
-    checkbox.addEventListener('change', ({ target }) => onChange(target))
-    checkbox.addEventListener('keydown', ({ target, key }) => {
-        if (key !== 'Enter') return;
-        onChange(target);
-    });
-}
-</script>
