--- src/components/landing/Callout.astro
@@ -1,31 +1,0 @@
----
-import Panel from '../Panel.astro';
-import Checklist from '../Checklist.astro';
----
-
-<section>
-    <Panel background="var(--gradient-pop-3)">
-        <Fragment slot="title">
-            <h3 class="head-md">Batteries Included.</h3>
-            <p>Astro does the heavy lifting in the background&mdash;no plugins required.</p>
-        </Fragment>
-
-        <Checklist>
-            <li>Markdown Support</li>
-            <li>TypeScript</li>
-            <li>CSS Modules</li>
-            <li>PostCSS, Sass, Less, Stylus</li>
-            <li>Built-in SEO</li>
-            <li>Sitemaps</li>
-            <li>RSS Feeds</li>
-            <li>Pagination</li>
-        </Checklist>
-    </Panel>
-</section>
-
-<style>
-    section {
-        background: var(--color-tan);
-        padding-bottom: 20rem;
-    }
-</style>
