+++ src/layouts/Base.astro
@@ -1,0 +1,57 @@
+---
+import { Sprite } from 'astro-icon';
+import BaseHead from '../components/BaseHead.astro';
+import SkipLink from '../components/SkipLink.astro';
+import type { Props as BaseHeadProps } from '../components/BaseHead.astro';
+
+export interface Props {
+    wrapper?: any;
+    meta: BaseHeadProps;
+    container?: boolean;
+}
+
+const { wrapper: Wrapper = Fragment, meta, container = false } = Astro.props;
+---
+
+<html lang="en">
+  <head>
+    <BaseHead {...meta} />
+
+    <slot name="head" />
+
+    <script>
+        document.body.classList.add('js');
+    </script>
+  </head>
+  <body>
+    <Sprite.Provider>
+        <slot name="banner" />
+
+        <SkipLink />
+        <slot name="nav" />
+
+        <Wrapper>
+            <slot name="start" />
+            <main id="content" class={container ? "container": ""}>
+                <slot />
+            </main>
+        </Wrapper>
+
+        <slot name="footer" />
+    </Sprite.Provider>
+
+    <script type="module" hoist>
+        import "/src/scripts/defer.ts";
+    </script>
+  </body>
+</html>
+
+<style global>
+    :root {
+        scroll-padding-top: 10rem;
+    }
+    main {
+        max-width: 100vw;
+        overflow: hidden;
+    }
+</style>
