+++ src/layouts/Page.astro
@@ -1,49 +1,33 @@
 ---
-import BaseHead from '../components/BaseHead.astro';
-import BaseLayout from '../components/BaseLayout.astro';
+import Layout from './Base.astro';
 import Nav from '../components/Nav.astro';
 import Footer from '../components/Footer.astro';
+import type { Props as BaseHeadProps } from '../components/BaseHead.astro';
 
-import { smartypants } from 'smartypants';
-
-let { title, description } = Astro.props;
-title = smartypants(title, 1);
-const meta = {
-	title: `${title} | Astro`,
-	description,
-	canonicalURL: Astro.request.canonicalURL
+export interface Props {
+    meta: BaseHeadProps;
 }
+
+const { meta } = Astro.props as Props;
 ---
 
-<html lang="en">
+<Layout container meta={meta}>
+    <Nav slot="nav" />
 
-<head>
-	<BaseHead {...meta} />
-	<slot name="head" />
+    <slot />
 
-	<style>
-		:root {
-			scroll-padding-top: 10rem;
-			background: var(--color-tan);
-		}
-		:global(main) {
-			padding-bottom: 8rem;
-			min-height: 100vh;
-			display: flex;
-			flex-direction: column;
-			align-items: center;
-		}
-	</style>
-</head>
+    <Footer slot="footer" />
+</Layout>
 
-<body>
-	<BaseLayout container>
-		<Nav slot="nav" />
-
-		<slot />
-
-		<Footer slot="footer" />
-	</BaseLayout>
-</body>
-
-</html>
+<style global>
+    :root {
+        background: var(--color-tan);
+    }
+    main {
+        padding-bottom: 8rem;
+        min-height: 100vh;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+</style>
